Use database as queue driver
Using cloudtasks to handle queue is not very efficient because the task
handler is still back to the laravel app, and GPT activity is mostly
blocked by waiting for reply and this could block other user-facing
requests.

So we add the "php artisan queue:work" process alongside the main
laravel app using supervisord, and change the QUEUE_CONNECTION to
"database".

* Add additional-supervisord.conf, which will make App Engine use
  supervisord to manage processes in the docker container, and specify
  the additional process as "php artisan queue:work".
* Rename docker-compose.yml.generated -> docker-compose.generated.yml
  for better docker-compose autocomplete.

Change-Id: Ia6f8ab27a2d2329781229232a90b0da94c9943d7