Add patch to cast enum from array
MongoDB may store enum as ['name' => '', 'value' => ''] instead of just
the string value. This happens in GCP runtime probably because the
MongoDB driver is old 1.12 and has not been updated to take care of
BackedEnum that is a recent PHP feature. While waiting for GCP runtime
to update MongoDB driver to at least 1.16, this patch is the workaround
to prevent errors.

Change-Id: I74bbe080a2280715418c994d555ff339fe9626c3